# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  version: 0.0.1
servers: 
  - url: http://localhost:3000
...
paths: 
...
  /users/{username}/photos/:
    parameters:
    - $ref: '#/components/parameters/username'
    - $ref: '#/components/parameters/page'
    - $ref: '#/components/parameters/limit'
    get: 
      operationId: getPhotos
      summary: Get the photos uploaded by the user. 
      description: Allows the user to get photos of the user specified by username.
      tags: 
        - photos
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photos" } 
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post: 
      operationId: uploadPhoto
      summary: Post a photo of the user and so add it to their stream of photos.
      description: Allows a user specified by username to add a photo to their photos collection. 
      tags: 
        - photos
      requestBody: 
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/Image" }
        required: true
      responses:
        '201':
          description: Photo uploaded successfully.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" } 
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /users/{username}/photos/{photoId}:
    parameters:
    - $ref: '#/components/parameters/username'
    - $ref: '#/components/parameters/photoId'
    delete: 
      operationId: deletePhoto
      summary: Delete a photo of the user and so remove it from their stream of photos.
      description: Allows the user to delete the photo specified by photoID of the user specified by username.
      tags: 
        - photos
      responses:
        '204':  {description: Photo deleted successfully. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    get: 
      operationId: getPhoto
      summary: Get the photo with the specified ID. 
      description: Allows the user to get photo of the user specified by username.
      tags: 
        - photos
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Photo" } 
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /users/{username}/photos/{photoId}/likes:
    parameters:
    - $ref: '#/components/parameters/username'
    - $ref: '#/components/parameters/photoId'
    - $ref: '#/components/parameters/page'
    - $ref: '#/components/parameters/limit'
    get: 
      operationId: getLikes
      summary: Get likes of a photo.
      description: A user can get the likes on a photo specified by photoId of a user username.
      tags: 
        - likes
      responses: 
        "200":
          description: Successful operation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Likes" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" } 
...
  /users/{username}/followers/:
    parameters:
    - $ref: '#/components/parameters/username'
    - $ref: '#/components/parameters/page'
    - $ref: '#/components/parameters/limit'
    get: 
      operationId: getFollowers
      summary: Get followers of a user specified by username.
      description: A user can get the followers of the user specified by username. 
      tags: 
        - followers
      responses: 
        "200":
          description: Successful operation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Followers" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }  
  /users/{username}/followings/:
    parameters:
    - $ref: '#/components/parameters/username'
    - $ref: '#/components/parameters/page'
    - $ref: '#/components/parameters/limit'
    get: 
      operationId: getFollowings
      summary: Get followings of a user specified by username.
      description: A user can get the users they are followed by the user specified by username. 
      tags: 
        - followings
      responses: 
        "200":
          description: Successful operation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Followings" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }  
...
  /users/{username}/bans/:
    parameters:
    - $ref: '#/components/parameters/username'
    - $ref: '#/components/parameters/page'
    - $ref: '#/components/parameters/limit'
    get:
      operationId: getBans
      summary: Get all users banned by the user username.
      description: The user specified by username can get all the bans they have put. 
      tags: 
        - bans
      responses: 
        "200":
          description: Successful operation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bans" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" } 
...
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user 
      description: |-
            If the user does not exist, it will be created,
            and an identifier is returned.
            If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json: 
            schema: { $ref: "#/components/schemas/Username" }
        required: true
      responses:
        '201':
          description: User log-in action successful 
          content:
            application/json: 
              schema: { $ref: "#/components/schemas/ID" }
components:
  securitySchemes:
    bearerAuth:   
      type: http
      scheme: bearer
      bearerFormat: JWT 
  parameters:
...
    username: 
      in: path
      name: username 
      description: The name of a user
      schema: {$ref: '#/components/schemas/Username'}
      required: true
...
    photoId:
      in: path
      name: photoId
      description: The identifier of a phot
      schema: {$ref: '#/components/schemas/PhotoId'}
      required: true
    page:
      name: page
      in: query
      description: Current page number
      required: false
      schema: 
        type: integer
        description: An integer signifying the page number
        format: int32
        default: 1
        minimum: 1
    limit:
      name: limit
      in: query
      description: Limits the number of returned items
      required: false
      schema: 
        type: integer
        description: An integer signifying the limit
        format: int32
        default: 10
        minimum: 10
        maximum: 100
        multipleOf: 10
  responses:
    BadRequest:
      description: Request was formulated badly
      content: 
        application/json:
          schema: 
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: There is an error in the server
      content: 
        application/json:
          schema: 
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: The resource was not found
      content: 
        application/json:
          schema: 
            $ref: '#/components/schemas/Error'
  schemas:
...
    Image: 
      description: The binary file corrisponding to the image
      type: string 
      format: binary
      minLength: 1
      maxLength: 100000
    PhotoId:
      description: The identifier for a photo assigned by the server
      type: integer
      example: 1
      minimum: 0 
      maximum: 1000
    ID: 
      description: The identifier for a user assigned by the server upon registration
      type: integer
      example: 1
      minimum: 0 
      maximum: 1000
...
    Username: 
      description: The name chosen by the user for public identification by other users
      type: string 
      example: Maria 
      pattern: '^[a-zA-Z][a-zA-Z0-9_.]{4,24}$'
      minLength: 5
      maxLength: 25
    Error:
      type: object 
      description: The error statement
      properties:
        statement:
          description: The text from the server for explaining the error
          type: string
          example: Invalid username.
          pattern: '^[a-zA-Z0-9_.+\- ,!?"();.]{5,30}$'
          minLength: 5
          maxLength: 50
    Photos: 
      type: array
      description: User's photos in reverse chronological order
      items: { $ref: "#/components/schemas/Photo" }
      minItems: 0
      maxItems: 1000
      uniqueItems: true
    Photo: 
      type: object
      description: The photo includes the id with which one can identify the photo and the image and the time posted as well as the likes and comments put.
      properties:
        ID: { $ref: "#/components/schemas/PhotoId" }
        Owner: { $ref: "#/components/schemas/Username" }
        Datetime: 
          description: The date of publishment of the photo
          type: string
          format: date-time
        NumLikes: 
          description: The number of likes of the photo
          type: integer
          format: int32
          example: 1
          minimum: 0 
          maximum: 1000
        NumComments: 
          description: The number of comments of the photo
          type: integer
          format: int32
          example: 1
          minimum: 0 
          maximum: 1000
      required: 
      - ID
      - Datetime
      - Owner
      - NumLikes
      - NumComments
    Follower:
      type: object
      description: A follower is reppresented by its username
      properties:
        Username: { $ref: "#/components/schemas/Username" }
      required: 
      - Username
    Followers:
      type: array
      description: The list of followers
      items: { $ref: "#/components/schemas/Follower" }
      minItems: 0
      maxItems: 1000
      uniqueItems: true
    Following:
      type: object
      description: The following user is reppresented by its username
      properties:
        Username: { $ref: "#/components/schemas/Username" }
      required: 
      - Username
    Followings:
      type: array
      description: The list of users followed
      items: { $ref: "#/components/schemas/Following" }
      minItems: 0
      maxItems: 1000
      uniqueItems: true
    Ban:
      type: object
      description: A banned user is reppresented by its username
      properties:
        Username: { $ref: "#/components/schemas/Username" }
      required: 
      - Username
    Bans:
      description: The list users banned
      type: array
      items: { $ref: "#/components/schemas/Ban" }
      minItems: 0
      maxItems: 1000
      uniqueItems: true
...
    Like: 
      type: object
      description: The user who put the like
      properties: 
        Username: { $ref: "#/components/schemas/Username" }    
      required: 
      - Username
    Likes:
      type: array
      description: The list of users who liked
      items: { $ref: "#/components/schemas/Like" }
      minItems: 0
      maxItems: 1000
      uniqueItems: true
...
security:
  - bearerAuth: []   